requirements.txt
# Requirements file for Infrastructure Damage Detection System

# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0

# Computer Vision
ultralytics>=8.0.211
opencv-python>=4.8.1.78
torch>=2.1.1
torchvision>=0.16.1
numpy>=1.24.3
scikit-image>=0.22.0

# Image processing
Pillow>=10.0.1

# Database
alembic>=1.13.0

# External APIs
twilio>=9.0.4
sendgrid>=6.11.0

# Utilities
requests>=2.31.0
aiofiles>=23.2.1

# Google Maps
# google-maps-services>=1.16.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1

# Security
python-multipart>=0.0.6
passlib>=1.7.4
python-jose>=3.3.0

# Development
jupyter>=1.0.0
ipython>=8.18.1
black>=23.11.0
flake8>=6.1.0
pylint>=3.0.3

# Production
gunicorn>=21.2.0
